\name{knock_out_two}
\alias{knock_out_two}
\title{Displays fixture for the tournament when number of teams is power of two}
\usage{
knock_out_two(number_of_teams)
}
\arguments{
  \item{number_of_teams}{input numeric}
     }
\value{
    \item{text}{Displays fixture for the tournament}
  }
\description{
Displays fixture for the tournament when number of teams is power of two
}
\examples{
number_of_teams=4
knock_out_two(number_of_teams)
}
